Add allow_group_by parameter for result grouping
If this is supplied it should contain a list of field names.  We will
look for good groups of results to return, based on the values of the
named fields.

This API is deliberately designed so that it doesn't guarantee that
groups will be returned, or allow control over the size of the groups or
way in which they're formed, since we expect this to require lots of
tuning based on behaviour of the system and of users.  The frontend
shouldn't need to worry about this tuning; it should just display the
groups that rummager returns.  Therefore, this tuning should be done in
rummager.