Add code to parse field_types.json
This commit adds a @field_types member to SchemaConfig, containing a
FieldTypes object which holds the parsed representation of the possible
field types.  Errors are raised at startup if the field_types.json file
is invalid in any way (eg, invalid JSON, unknown keys in a type
definition, no elasticsearch config defined for a type).