When indexing, fetch tags from the content API
This commit adjusts rummager's indexing process such that it makes a
call to the content-api (https://github.com/alphagov/govuk_content_api)
when the document is being prepared for sending to elasticsearch, and
uses information from that to populate the tagging fields (ie,
`organisations`, `mainstream_browse_pages`, and `specialist_sectors`).

This is, perhaps surprisingly, part of the migration away from using
content-api / panopticon for getting tag information.

The idea is that instead of applications having to send content to
rummager via panopticon, and using panopticon to add the tag
information, applications can instead send content directly to rummager,
and rummager will sort out getting the right tag information.  This
allows us to change publishing apps such that they no longer talk to
panopticon, while maintaining the correct data in rummager.

Once all the tag data has been populated in the content-store, and we've
got the content-store keeping sync with panopticon, we'll change this to
fetch tags from the content store instead of content-api.

Trello: https://trello.com/c/xQtX24kW