Remove bad document from test data
The test data for the sitemap had a "bad" document which doesn't have a
"link" field.  This is not something which should ever occur in our
search index, since we're using the link field as the unique identifier
for documents.  This document isn't actually used in any of the tests,
and removing it doesn't make the test fail.

Therefore, I'm removing this to make things simpler, and also because
doing so will allow me to use a generic helper method for adding test
documents in a later commit.