Stub the contentapi in integration test helpers
We now make a call to the content API whenever we add a document, so the
integration tests need to have this call stubbed out.  The easiest way
to do this is by using the `insert_document` test helper which many of
the tests already use, so this commit adds stubbing to that method.

It would be good to be able to call this test helper from more of the
integration tests, so this commit moves it from the MultiIndexTest class
to the IntegrationTest class.  I've also moved the `commit_document` and
`commit_index` methods, since they're also likely to be useful from
other tests, and exposed a `prepare_document_lookups` method for setting
up the stubs without adding a document, which will be useful for some of
the integration tests which check things like bulk indexing.