Use helpers to stub integration test requests
Change all integration tests to use the `prepare_document_lookups`
and `insert_document` helpers instead of direct calls, so that they stub
calls to the contentapi appropriately.  Doing it this way also means it
will be much easier to adjust that stubbing when we remove the call to
contentapi in future, and add a call to the content-store instead.