Add a downloader object to fetch and merge alpha taxonomy definitions
To support prototyping of a new tagging interface, it would be useful to
have API-based access to the taxonomy definitions that are currently
being put together by Information Architects in the Finding Things team.

This commit adds an object that fetches, processes and merges taxonomy
definitions stored in Google Drive. It also includes a Rake task to
run the download. This is part of a wider piece to import these taxonomy
definitions into a new field on documents indexed by Rummager, allowing
us to filter on these documents via the unified search api.